requirements.txt
aiohttp
curl_cffi
fastapi
httpx
jinja2
markdown2[all]
openai
pycrypto
pydantic
requests
sse_starlette
termcolor
tclogger
tiktoken
transformers
uvicorn
websockets